@@ -1,4 +1,4 @@
-import 'dart:collection';
+ import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:ytsync/network.dart';
 import 'package:ytsync/pages/homepage_widget.dart';
@@ -7,6 +7,11 @@ import './settings.dart';
 import 'announcement.dart';
 import '../util.dart';
 import '../main.dart';
+import 'package:firebase_auth/firebase_auth.dart'; 
+
+final FirebaseAuth _auth = FirebaseAuth.instance;
+bool _hoverRefresh = false;
+bool _hoverFilter = false;
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -14,16 +19,12 @@ class HomePage extends StatefulWidget {
   @override
   HomePageState createState() => HomePageState();
 }
-
 class HomePageState extends State<HomePage> {
-  //announcements and classes
   late List<AnnouncementData> announcements;
   List<String> selectedClasses = [];
   List<String> availableClasses = [];
-  HashMap<String, String> displayClasses =
-      HashMap<String, String>(); //actual name, display name
+  HashMap<String, String> displayClasses = HashMap<String, String>();
 
-  //filter category
   bool showCompleted = true;
   bool showUncompleted = true;
   bool showPersonal = true;
@@ -43,13 +44,12 @@ class HomePageState extends State<HomePage> {
       var selected = entry.selected;
 
       availableClasses.add(clazz);
-      displayClasses[clazz] =
-          clazz
-              .replaceAll("Sec 4", "")
-              .replaceAll("Sec 3", "")
-              .replaceAll("Sec 2", "")
-              .replaceAll("Sec 1", "")
-              .trimLeft();
+      displayClasses[clazz] = clazz
+          .replaceAll("Sec 4", "")
+          .replaceAll("Sec 3", "")
+          .replaceAll("Sec 2", "")
+          .replaceAll("Sec 1", "")
+          .trimLeft();
       if (selected) {
         selectedClasses.add(clazz);
       }
@@ -60,11 +60,8 @@ class HomePageState extends State<HomePage> {
 
   void _toggleClassSelection(String className, bool? value) {
     setState(() {
-      if (value == true) {
-        selectedClasses.add(className);
-      } else {
-        selectedClasses.remove(className);
-      }
+      if (value == true) selectedClasses.add(className);
+      else selectedClasses.remove(className);
     });
   }
 
@@ -79,7 +76,6 @@ class HomePageState extends State<HomePage> {
     String clazz,
     DateTime due,
     DateTime? publish,
-    String author,
     String description,
     String uuid,
     bool isPublic,
@@ -89,15 +85,12 @@ class HomePageState extends State<HomePage> {
       clazz,
       due,
       publish,
-      author,
       description,
       uuid,
       isPublic,
     );
 
-    if (!await sendAnnouncementToServer(data, isPublic)) {
-      return false;
-    }
+    if (!await sendAnnouncementToServer(data, isPublic)) return false;
 
     setState(() {
       announcements.add(data);
@@ -113,137 +106,232 @@ class HomePageState extends State<HomePage> {
     });
   }
 
-  MaterialPageRoute getSettingsPageMaterial() {
-    return MaterialPageRoute(
-      builder:
-          (context) => SettingsPage(
-            availableClasses: availableClasses,
-            selectedClasses: selectedClasses,
-            displayClasses: displayClasses,
-            onClassToggle: _toggleClassSelection,
-            onMultipleClassSelect: _setSelectedClasses,
-          ),
-    );
-  }
+MaterialPageRoute getSettingsPageMaterial() {
+  return MaterialPageRoute(
+    builder: (context) => SettingsPage(
+      selectedClasses: selectedClasses,
+      displayClasses: displayClasses,
+      onClassToggle: _toggleClassSelection,
+      onMultipleClassSelect: _setSelectedClasses,
+    ),
+  );
+}
 
-  void showFilterPopup(context) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        final theme = Theme.of(context);
-        return AlertDialog(
-          scrollable: true,
-          title: const Text("Filter Announcements"),
-          titleTextStyle: theme.textTheme.titleMedium?.copyWith(
-            fontWeight: FontWeight.bold,
-            letterSpacing: 1.5,
-            fontSize: 30.0,
+  void showFilterPopup(BuildContext context) {
+    final theme = Theme.of(context);
+
+    final size = MediaQuery.of(context).size;
+    final bool isPhone = size.width < 600;
+
+    Color color1 = const Color.fromARGB(255, 17, 6, 161);
+    Color color2 = const Color(0xFFFFC700);
+    Color color3 = Colors.white;
+
+    if (appState.selectedTheme == 'light') {
+      color1 = const Color(0xFF0070D1);
+      color2 = Colors.grey.shade200;
+      color3 = Colors.white;
+    } else if (appState.selectedTheme == 'dark') {
+      color1 = Colors.white;
+      color2 = Colors.grey.shade900;
+      color3 = Colors.grey.shade800;
+    }
+
+showDialog(
+  context: context,
+  builder: (BuildContext context) {
+    return Dialog(
+      backgroundColor: Colors.transparent,
+      insetPadding: const EdgeInsets.all(16),
+      child: Material(
+        borderRadius: BorderRadius.circular(20),
+        color: color3,
+        elevation: 8,
+        child: Container(
+          width: 350,
+          padding: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+            color: color3,
+            borderRadius: BorderRadius.circular(20),
+            border: Border.all(
+              color: Colors.grey.shade500,
+              width: 2,
+            ),
           ),
-          content: StatefulBuilder(
-            builder:
-                (context, setState) => Padding(
-                  padding: const EdgeInsets.all(12.0),
-                  child: Form(
-                    child: Table(
-                      columnWidths: {
-                        0: FlexColumnWidth(),
-                        1: FlexColumnWidth(0.2),
-                      },
+          child: StatefulBuilder(
+            builder: (context, setState) => Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                Text(
+                  "Filter Announcements",
+                  style: TextStyle(
+                    fontFamily: "montserrat",
+                    fontSize: 22,
+                    fontWeight: FontWeight.bold,
+                    color: color1,
+                  ),
+                  textAlign: TextAlign.center,
+                ),
+                const SizedBox(height: 20),
+                Column(
+                  children: [
+                    // Show Completed
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        TableRow(
-                          children: [
-                            Text(
-                              "Show Completed",
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 20.0,
-                              ),
+                        const Expanded(
+                          child: Text(
+                            "Show Completed",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
                             ),
-                            Checkbox(
-                              checkColor: Colors.white,
-                              value: showCompleted,
-                              onChanged: (bool? value) {
-                                setState(() {
-                                  showCompleted = value!;
-                                });
-                              },
-                            ),
-                          ],
+                          ),
                         ),
-                        TableRow(
-                          children: [
-                            Text(
-                              "Show Incomplete",
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 20.0,
-                              ),
-                            ),
-                            Checkbox(
-                              checkColor: Colors.white,
-                              value: showUncompleted,
-                              onChanged: (bool? value) {
-                                setState(() {
-                                  showUncompleted = value!;
-                                });
-                              },
-                            ),
-                          ],
-                        ),
-                        TableRow(
-                          children: [
-                            Text(
-                              "Show Personal",
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 20.0,
-                              ),
-                            ),
-                            Checkbox(
-                              checkColor: Colors.white,
-                              value: showPersonal,
-                              onChanged: (bool? value) {
-                                setState(() {
-                                  showPersonal = value!;
-                                });
-                              },
-                            ),
-                          ],
-                        ),
-                        TableRow(
-                          children: [
-                            Text(
-                              "Show Public",
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 20.0,
-                              ),
-                            ),
-                            Checkbox(
-                              checkColor: Colors.white,
-                              value: showPublic,
-                              onChanged: (bool? value) {
-                                setState(() {
-                                  showPublic = value!;
-                                });
-                              },
-                            ),
-                          ],
+                        Checkbox(
+                          value: showCompleted,
+                          onChanged: (bool? value) {
+                            setState(() {
+                              showCompleted = value ?? false;
+                            });
+                          },
+                          activeColor: color1,
+                          checkColor: appState.selectedTheme == 'dark'
+                              ? Colors.black
+                              : Colors.white,
+                          side: BorderSide(color: color1, width: 2),
                         ),
                       ],
                     ),
-                  ),
+                    const SizedBox(height: 8),
+
+                    // Show Incomplete
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        const Expanded(
+                          child: Text(
+                            "Show Incomplete",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
+                            ),
+                          ),
+                        ),
+                        Checkbox(
+                          value: showUncompleted,
+                          onChanged: (bool? value) {
+                            setState(() {
+                              showUncompleted = value ?? false;
+                            });
+                          },
+                          activeColor: color1,
+                          checkColor: appState.selectedTheme == 'dark'
+                              ? Colors.black
+                              : Colors.white,
+                          side: BorderSide(color: color1, width: 2),
+                        ),
+                      ],
+                    ),
+                    const SizedBox(height: 8),
+
+                    // Show Personal
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        const Expanded(
+                          child: Text(
+                            "Show Personal",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
+                            ),
+                          ),
+                        ),
+                        Checkbox(
+                          value: showPersonal,
+                          onChanged: (bool? value) {
+                            setState(() {
+                              showPersonal = value ?? false;
+                            });
+                          },
+                          activeColor: color1,
+                          checkColor: appState.selectedTheme == 'dark'
+                              ? Colors.black
+                              : Colors.white,
+                          side: BorderSide(color: color1, width: 2),
+                        ),
+                      ],
+                    ),
+                    const SizedBox(height: 8),
+
+                    // Show Public
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        const Expanded(
+                          child: Text(
+                            "Show Public",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
+                            ),
+                          ),
+                        ),
+                        Checkbox(
+                          value: showPublic,
+                          onChanged: (bool? value) {
+                            setState(() {
+                              showPublic = value ?? false;
+                            });
+                          },
+                          activeColor: color1,
+                          checkColor: appState.selectedTheme == 'dark'
+                              ? Colors.black
+                              : Colors.white,
+                          side: BorderSide(color: color1, width: 2),
+                        ),
+                      ],
+                    ),
+                  ],
                 ),
-          ),
-          actions: [
-            ElevatedButton(
-              child: const Text("Filter"),
-              onPressed: () {
-                setState(() {});
-                Navigator.pop(context);
-              },
+                    const SizedBox(height: 25),
+                    ElevatedButton(
+                      style: ElevatedButton.styleFrom(
+                        backgroundColor: appState.selectedTheme == 'light'
+                            ? const Color(0xFF0070D1)
+                            : appState.selectedTheme == 'dark'
+                            ? Colors.white
+                            : color2,
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        padding: const EdgeInsets.symmetric(vertical: 14, horizontal: 30),
+                      ),
+                      onPressed: () {
+                        setState(() {});
+                        Navigator.pop(context);
+                      },
+                      child: Text(
+                        "Apply Filter",
+                        style: TextStyle(
+                          fontFamily: "montserrat",
+                          fontWeight: FontWeight.bold,
+                          fontSize: 16,
+                          color: appState.selectedTheme == 'light'? Colors.white : Colors.black,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
             ),
-          ],
+          ),
         );
       },
     );
@@ -252,6 +340,29 @@ class HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+
+    final size = MediaQuery.of(context).size;
+    final bool isPhone = size.width < 600;
+
+    Color color1 = const Color.fromARGB(255, 17, 6, 161);
+    Color color2 = const Color(0xFFFFC700);
+    Color color3 = Colors.white;
+    String filterAsset = "assets/filterdark.png";
+    String plusAsset = "assets/plus.png";
+    String gearAsset = "assets/gear.png";
+
+    if (appState.selectedTheme == 'light') {
+      color1 = const Color(0xFF0070D1);
+      color2 = Colors.grey.shade200;
+      color3 = Colors.white;
+    } else if (appState.selectedTheme == 'dark') {
+      color1 = Colors.white;
+      color2 = Colors.grey.shade900;
+      color3 = Colors.grey.shade800;
+      filterAsset = "assets/filter.png";
+      plusAsset = "assets/plusdark.png";
+      gearAsset = "assets/geardark.png";
+    }
 
     refreshFunc() async {
       selectedClasses.clear();
@@ -263,84 +374,211 @@ class HomePageState extends State<HomePage> {
 
       setState(() {
         homepageInit();
-
         showSnackBar(context, "Refreshed Page.");
       });
     }
 
     return Scaffold(
+      backgroundColor: color2,
       appBar: AppBar(
+        backgroundColor: color1,
         automaticallyImplyLeading: false,
-        title: Text(
-          "Announcements",
-          style: TextStyle(fontWeight: FontWeight.bold),
-        ),
-      ),
-
-      body: Padding(
-        padding: EdgeInsets.all(16.0),
-        child: Column(
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Text("Today is ${fullDateStr(DateTime.now())}", style: theme.textTheme.bodyMedium?.copyWith(
-              fontWeight: FontWeight.bold
-            )),
+            // LEFT: Profile
             Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                IconButton(
-                  icon: Icon(Icons.filter_alt_outlined),
-                  tooltip: "Pick a filter to categorise the announcements.",
-                  onPressed: () => showFilterPopup(context),
+                CircleAvatar(
+                  radius: 18,
+                  backgroundColor: appState.selectedTheme == 'dark'
+                      ? Colors.black
+                      : Colors.white,
+                  child: Icon(
+                    Icons.person,
+                    size: 20,
+                    color: appState.selectedTheme == 'dark'
+                        ? Colors.white
+                        : color1,
+                  ),
                 ),
-                Row(
+                const SizedBox(width: 10),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    Icon(Icons.refresh_outlined),
-                    TimerButton.builder(
-                      onPressed: refreshFunc,
-                      builder: (context, timeLeft) {
-                        return MouseRegion(
-                          cursor:
-                              timeLeft >= 0
-                                  ? DefaultMouseCursor()
-                                  : SystemMouseCursors.click,
-                          child:
-                              timeLeft >= 0
-                                  ? Text(
-                                    "Refresh (${timeLeft}s)",
-                                    style: theme.textTheme.labelLarge,
-                                  )
-                                  : Text(
-                                    "Refresh",
-                                    style: theme.textTheme.labelLarge?.copyWith(
-                                      color:
-                                          appState
-                                              .themeData
-                                              .$2
-                                              .colorScheme
-                                              .secondary,
-                                    ),
-                                  ),
-                        );
-                      },
-                      timeOutInSeconds: 60,
+                    Text(
+                      account.name,
+                      style: TextStyle(
+                        fontFamily: "montserrat",
+                        fontWeight: FontWeight.bold,
+                        fontSize: 14,
+                        color: appState.selectedTheme == 'dark'
+                            ? Colors.black
+                            : Colors.white,
+                      ),
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                    const SizedBox(height: 2),
+                    Text(
+                      account.email,
+                      style: TextStyle(
+                        fontFamily: "montserrat",
+                        fontSize: 12,
+                        color: appState.selectedTheme == 'dark'
+                            ? Colors.black.withOpacity(0.8)
+                            : Colors.white.withOpacity(0.9),
+                      ),
                     ),
                   ],
                 ),
               ],
             ),
-
-            Divider(),
-
-            Expanded(
-              child:
-                  announcements.isEmpty
-                      ? Center(
-                        child: Text(
-                          "No announcements to display",
-                          style: theme.textTheme.titleMedium,
+          ],
+        ),
+      ),
+      body: Column(
+        children: [
+          // Top bar with filter and refresh
+          Container(
+            width: double.infinity,
+            color: color2,
+            padding: EdgeInsets.fromLTRB(
+              16,
+              isPhone ? 10 : 20, 
+              16,
+              isPhone ? 10 : 20, 
+            ),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Row(
+                  children: [
+                    // LEFT: Filter Button
+                    Expanded(
+                      child: Align(
+                        alignment: Alignment.centerLeft,
+                        child: MouseRegion(
+                          onEnter: (_) => setState(() => _hoverFilter = true),
+                          onExit: (_) => setState(() => _hoverFilter = false),
+                          cursor: SystemMouseCursors.click,
+                          child: GestureDetector(
+                            onTap: () => showFilterPopup(context),
+                            child: AnimatedContainer(
+                              duration: const Duration(milliseconds: 200),
+                              padding: EdgeInsets.all(isPhone ? 4 : 8),
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(12),
+                                border: Border.all(
+                                  color: _hoverFilter ? color1 : Colors.transparent,
+                                  width: 2,
+                                ),
+                              ),
+                              child: Image.asset(
+                                filterAsset,
+                                width: isPhone ? 26 : 28,  
+                                height: isPhone ? 26 : 28,
+                              ),
+                            ),
+                          ),
                         ),
-                      )
-                      : (createAnnouncementList(
+                      ),
+                    ),
+
+                    // RIGHT: Refresh Button
+                    Expanded(
+                      child: Align(
+                        alignment: Alignment.centerRight,
+                        child: MouseRegion(
+                          onEnter: (_) => setState(() => _hoverRefresh = true),
+                          onExit: (_) => setState(() => _hoverRefresh = false),
+                          cursor: SystemMouseCursors.click,
+                          child: GestureDetector(
+                            onTap: refreshFunc,
+                            child: AnimatedContainer(
+                              duration: const Duration(milliseconds: 200),
+                              padding: EdgeInsets.symmetric(
+                                horizontal: isPhone ? 6 : 8,
+                                vertical: isPhone ? 2 : 4,
+                              ),
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(12),
+                                border: Border.all(
+                                  color: _hoverRefresh ? color1 : Colors.transparent,
+                                  width: 2,
+                                ),
+                              ),
+                              child: Row(
+                                mainAxisSize: MainAxisSize.min,
+                                children: [
+                                  Icon(
+                                    Icons.refresh_outlined,
+                                    size: isPhone ? 26 : 28,
+                                    color: appState.selectedTheme == 'dark'
+                                        ? Colors.white
+                                        : Colors.black87,
+                                  ),
+                                  const SizedBox(width: 4),
+                                  Text(
+                                    "Refresh",
+                                    style: theme.textTheme.labelLarge?.copyWith(
+                                      fontFamily: "montserrat",
+                                      fontSize: isPhone ? 13 : 16,
+                                      color: appState.selectedTheme == 'dark'
+                                          ? Colors.white
+                                          : Colors.black87,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+                SizedBox(height: isPhone ? 6 : 10),
+                Divider(
+                  thickness: 1,
+                  color: appState.selectedTheme == 'dark'
+                      ? Colors.white54
+                      : Colors.black54,
+                ),
+              ],
+            ),
+          ),
+
+          // Main announcement list
+          Expanded(
+            child: Container(
+              width: double.infinity,
+              margin: EdgeInsets.symmetric(
+                horizontal: 10,
+                vertical: isPhone ? 0 : 8,
+              ),
+              decoration: BoxDecoration(
+                color: color3,
+                border: Border.all(
+                  color: color1,
+                  width: 3,
+                ),
+                borderRadius: BorderRadius.circular(12),
+              ),
+              child: announcements.isEmpty
+                  ? Center(
+                      child: Text(
+                        "No announcements to display",
+                        style: theme.textTheme.titleMedium?.copyWith(
+                          fontFamily: "montserrat",
+                          fontWeight: FontWeight.bold,
+                          color: color1,
+                        ),
+                      ),
+                    )
+                  : Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: createAnnouncementList(
                         announcements,
                         selectedClasses,
                         displayClasses,
@@ -352,70 +590,229 @@ class HomePageState extends State<HomePage> {
                         context,
                         setState,
                         removeAnnouncement,
-                      )),
+                      ),
+                    ),
+            ),
+          ),
+
+          // Bottom bar buttons
+          Container(
+            width: double.infinity,
+            color: color2,
+            padding: EdgeInsets.symmetric(
+              vertical: isPhone ? 10 : 20, 
+              horizontal: 16,
+            ),
+            child: Row(
+              children: [
+                Expanded(
+                  child: ElevatedButton(
+                    onPressed: () {
+                      if (selectedClasses.isNotEmpty) { 
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder:
+                                (context) => AddAnnouncementPage(
+                                  homePageState: this,
+                                  availableClasses: availableClasses,
+                                  selectedClasses: selectedClasses,
+                                  displayClasses: displayClasses,
+                                ),
+                          ),
+                        );
+                      } else {
+                        showDialog(
+                          context: context,
+                          builder: (BuildContext context) {
+                            return Dialog(
+                              backgroundColor: Colors.transparent,
+                              insetPadding: const EdgeInsets.all(16),
+                              child: Material(
+                                borderRadius: BorderRadius.circular(20),
+                                color: color3,
+                                elevation: 8,
+                                child: Container(
+                                  width: 360,
+                                  padding: const EdgeInsets.all(24),
+                                  decoration: BoxDecoration(
+                                    color: color3,
+                                    borderRadius: BorderRadius.circular(20),
+                                    border: Border.all(
+                                      color: Colors.grey.shade500,
+                                      width: 2,
+                                    ),
+                                  ),
+                                  child: Column(
+                                    mainAxisSize: MainAxisSize.min,
+                                    children: [
+                                      Text(
+                                        "No Classes Selected",
+                                        style: TextStyle(
+                                          fontFamily: "montserrat",
+                                          fontSize: 22,
+                                          fontWeight: FontWeight.bold,
+                                          color: color1,
+                                        ),
+                                        textAlign: TextAlign.center,
+                                      ),
+                                      const SizedBox(height: 16),
+                                      Text(
+                                        "Please select at least one class to continue.",
+                                        textAlign: TextAlign.center,
+                                        style: TextStyle(
+                                          fontFamily: "montserrat",
+                                          fontSize: 16,
+                                          color: appState.selectedTheme == 'dark'
+                                          ? Colors.white
+                                          : Colors.black,
+                                        ),
+                                      ),
+                                      const SizedBox(height: 24),
+                                      Row(
+                                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                                        children: [
+                                          OutlinedButton(
+                                            onPressed: () => Navigator.pop(context),
+                                            style: OutlinedButton.styleFrom(
+                                              side: BorderSide(color: Colors.grey.shade600, width: 2),
+                                              shape: RoundedRectangleBorder(
+                                                borderRadius: BorderRadius.circular(12),
+                                              ),
+                                              padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+                                            ),
+                                            child: Text(
+                                              "Cancel",
+                                              style: TextStyle(
+                                                fontFamily: "montserrat",
+                                                fontWeight: FontWeight.bold,
+                                                fontSize: 16,
+                                                color: appState.selectedTheme == 'dark'
+                                                ? Colors.white
+                                                : Colors.black,
+                                              ),
+                                            ),
+                                          ),
+                                          ElevatedButton(
+                                            onPressed: () {
+                                              Navigator.pop(context);
+                                              Navigator.push(context, getSettingsPageMaterial());
+                                            },
+                                            style: ElevatedButton.styleFrom(
+                                              backgroundColor: appState.selectedTheme == 'light'
+                                                ? const Color(0xFF0070D1)
+                                                : appState.selectedTheme == 'dark'
+                                                ? Colors.white
+                                                : color2,
+                                              shape: RoundedRectangleBorder(
+                                                borderRadius: BorderRadius.circular(12),
+                                              ),
+                                              padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+                                            ),
+                                            child: Text(
+                                              "Go to Settings",
+                                              style: TextStyle(
+                                                fontFamily: "montserrat",
+                                                fontWeight: FontWeight.bold,
+                                                fontSize: 16,
+                                                color: appState.selectedTheme == 'light'
+                                                    ? Colors.white
+                                                    : Colors.black,
+                                              ),
+                                            ),
+                                          ),
+                                        ],
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                                ),
+                              );
+                            },
+                          );
+                        }
+                      },
+                      style: ElevatedButton.styleFrom(
+                        backgroundColor: color1,
+                        padding: EdgeInsets.symmetric(
+                          vertical: isPhone ? 6 : 10, 
+                        ),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                      ),
+                      child: Column(
+                        mainAxisSize: MainAxisSize.min,
+                        children: [
+                          SizedBox(height: isPhone ? 2 : 6),
+                          Image.asset(
+                            plusAsset,
+                            width: isPhone ? 22 : 28, 
+                            height: isPhone ? 22 : 28,
+                          ),
+                          SizedBox(height: isPhone ? 2 : 6),
+                          Text(
+                            "Add Announcement",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: isPhone ? 10 : 12,
+                              color: appState.selectedTheme == 'dark'
+                                  ? Colors.black
+                                  : Colors.white,
+                            ),
+                          ),
+                          SizedBox(height: isPhone ? 2 : 6),
+                        ],
+                      ),
+                    ),
+                  ),
+                  const SizedBox(width: 10),
+                  Expanded(
+                    child: ElevatedButton(
+                      onPressed: () {
+                        Navigator.push(context, getSettingsPageMaterial());
+                      },
+                      style: ElevatedButton.styleFrom(
+                        backgroundColor: color1,
+                        padding: EdgeInsets.symmetric(
+                          vertical: isPhone ? 6 : 10,
+                        ),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                      ),
+                      child: Column(
+                        mainAxisSize: MainAxisSize.min,
+                        children: [
+                          SizedBox(height: isPhone ? 2 : 6),
+                          Image.asset(
+                            gearAsset,
+                            width: isPhone ? 22 : 28,
+                            height: isPhone ? 22 : 28,
+                          ),
+                          SizedBox(height: isPhone ? 2 : 6),
+                          Text(
+                            "Settings",
+                            style: TextStyle(
+                              fontFamily: "montserrat",
+                              fontWeight: FontWeight.bold,
+                              fontSize: isPhone ? 10 : 12,
+                              color: appState.selectedTheme == 'dark'
+                                  ? Colors.black
+                                  : Colors.white,
+                            ),
+                          ),
+                          SizedBox(height: isPhone ? 2 : 6),
+                        ],
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ],
         ),
-      ),
-
-      bottomNavigationBar: BottomNavigationBar(
-        items: [
-          BottomNavigationBarItem(icon: Icon(Icons.add), label: "Add"),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.settings),
-            label: "Settings",
-          ),
-        ],
-        onTap: (index) {
-          if (index == 0) {
-            // Add
-            if (selectedClasses.isNotEmpty) {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder:
-                      (context) => AddAnnouncementPage(
-                        homePageState: this,
-                        availableClasses: availableClasses,
-                        selectedClasses: selectedClasses,
-                        displayClasses: displayClasses,
-                      ),
-                ),
-              );
-            } else {
-              showDialog(
-                context: context,
-                builder:
-                    (context) => AlertDialog(
-                      title: Text("No classes selected."),
-                      content: Text(
-                        "Please select at least one class to continue.",
-                      ),
-                      actions: [
-                        TextButton(
-                          onPressed: () => Navigator.pop(context),
-                          child: Text("Cancel"),
-                        ),
-                        ElevatedButton(
-                          onPressed: () {
-                            Navigator.pop(context);
-                            Navigator.push(context, getSettingsPageMaterial());
-                          },
-                          child: Text("Go to Settings"),
-                        ),
-                      ],
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(16),
-                      ),
-                    ),
-              );
-            }
-          } else if (index == 1) {
-            // Settings
-            Navigator.push(context, getSettingsPageMaterial());
-          }
-        },
-      ),
     );
   }
 }
